refactor(projects): extract BaselineManagement component (990 lines)
- Extracted BaselineManagement from page.tsx to components/BaselineManagement.tsx
- Reduced main file from 4,969 to 3,975 lines (20% reduction!)
- All functionality preserved, zero linter errors
- Component includes baseline extraction, drift detection, approval workflow
- Part of File #4 refactoring initiative (CR-2026-001)
